Image Steganography Tool
Every artform contains a hidden message. Images is one of them. But what if your are actually able to hide messages and send the images with your containing messages, Emessages is the way out. This is a basic image steganography tool to encode and decode images with the text message.